purchase receipt + delivery note + proforma invoice print format , edit on custom stock blaance , ledger report , sales order taxes

+++ dmc/dmc/report/custom_stock_balance/custom_stock_balance.py
@@ -450,7 +450,7 @@ class StockBalanceReport:
                 {
                     "label": _("Balance Qty"),
                     "fieldname": "bal_qty",
-                    "fieldtype": "Float",
+                    "fieldtype": "Int",
                     "width": 100,
                     "convertible": "qty",
                 },
