+++ dmc/dmc/report/custom_stock_ledger/custom_stock_ledger.py
@@ -261,7 +261,7 @@ def get_columns(filters):
             {
                 "label": _("Balance Qty"),
                 "fieldname": "qty_after_transaction",
-                "fieldtype": "Float",
+                "fieldtype": "Int",
                 "width": 100,
                 "convertible": "qty",
             },
